// Console/RowConsole/packages/core/src/lib/offline/row-contact/models/row-contact.dto.ts
export interface RowContactDto {
    row_contact_unique_id: string;
    communication_detail?: string;
    communication_type?: string;
    system_contact_type?: string;
    system_contact_category?: string;
    system_contact_group?: string;
    system_contact_status?: string;
    dnd_status?: string;
    contact_person?: string;
    organization_name?: string;
    pin_code?: string;
    landmark?: string;
    zone?: string;
    country?: string;
    state?: string;
    city?: string;
    address?: string;
    latitude?: string;
    longitude?: string;
    create_date?: string; // ISO String
    verification_date?: string;
    next_verification_date?: string;
    last_activity_date_time?: string;
    campaign_code?: string;
    licence_reg_no_1?: string;
    licence_reg_no_2?: string;
    licence_reg_no_3?: string;
    licence_reg_no_4?: string;
    licence_reg_no_5?: string;
    organization_unique_id?: string;
    contact_unique_id?: string;
    row_data?: string;
    no_of_organization?: string;
    no_of_contact?: number;
    verification_status?: string;
    mapping_row_to_main_contact?: string;
    mapping_row_to_main_organization?: string;
    other_field_a?: string;
    other_field_b?: string;
    other_field_c?: string;
    other_field_d?: string;
    other_field_e?: string;
    other_field_f?: string;
    other_field_g?: string;
    other_field_h?: string;
    other_field_i?: string;
    other_field_j?: string;
    other_field_k?: string;
    other_field_l?: string;
    other_field_m?: string;
    mobile_number?: string;
    email_id?: string;
    product_detail_a?: string;
    product_detail_b?: string;
    product_detail_c?: string;
    config_json?: string;
    is_active?: string;
}